efi2025/requirements.txt
alembic==1.13.2
asgiref==3.8.1
asttokens==3.0.0
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
decorator==5.2.1
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
executing==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==23.0.0
inflection==0.5.1
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
matplotlib-inline==0.1.7
mysql-connector-python==9.0.0
mysqlclient==2.2.5
numpy==2.2.6
packaging==24.1
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
polib==1.2.0
prompt_toolkit==3.0.50
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
PyMySQL==1.1.1
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
shiboken6==6.6.0
six==1.17.0
SQLAlchemy==2.0.31
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
